Refactor namespaces from Models to Entities
This commit updates the namespace from `NotesApp.Domain.Models` to `NotesApp.Domain.Entities` across multiple files, including controllers, services, models, repositories, and configuration files. This change clarifies the purpose of the classes as entities. All relevant imports have been adjusted to ensure the application functions correctly with the new namespace.
