+++ WebApplication4/API/Controllers/GroupController.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using NotesApp.Application.DTOs;
 using NotesApp.Application.Services;
-using NotesApp.Domain.Models;
+using NotesApp.Domain.Entities;
 using NotesApp.Infrastructure.Data;
 
 namespace NotesApp.API.Controllers
