+++ WebApplication4/Domain/Entities/User.cs
@@ -1,6 +1,6 @@
 ﻿using NotesApp.Domain.Interfaces;
 
-namespace NotesApp.Domain.Models
+namespace NotesApp.Domain.Entities
 {
     public class User : StandardModel, ISoftDelete
     {
