+++ WebApplication4/Infrastructure/Repositories/UserRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using NotesApp.Application.DTOs;
+using NotesApp.Domain.Entities;
 using NotesApp.Domain.Interfaces;
-using NotesApp.Domain.Models;
 using NotesApp.Infrastructure.Data;
 
 namespace NotesApp.Infrastructure.Repositories
